% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.SpatialPolygonsDataFrame.spong}
\alias{as.SpatialPolygonsDataFrame.spong}
\title{spong to spolydf}
\usage{
as.SpatialPolygonsDataFrame.spong(nsp)
}
\arguments{
\item{nsp}{a spong}
}
\value{
a spatal poly df
}
\description{
spong to spolydf
}
\details{
.. content for \details{} ..
}
\author{
Barry Rowlingson
}

