% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.spong.SpatialPolygonsDataFrame}
\alias{as.spong.SpatialPolygonsDataFrame}
\title{convert spolydf to spong}
\usage{
as.spong.SpatialPolygonsDataFrame(spdf)
}
\arguments{
\item{spdf}{a spatial poly df}
}
\value{
a spong
}
\description{
convert spolydf to spong
}
\details{
.. content for \details{} ..
}
\author{
Barry Rowlingson
}

